Filter temperature series in get_temperature_series and only return requested range

+++ src/sources/common.rs
@@ -65,7 +65,7 @@ impl Thousandth for f64 {
 #[derive(Deserialize, Debug, Clone)]
 pub struct SingleTemperature {
     #[serde(rename = "time")]
-    pub epoch: u32,
+    pub epoch: i64,
     pub temperature_2m: f32,
 }
